pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-storage-blob>=12.28.0",
    "pydantic-settings>=2.13.1",
]
